requirements_enterprise.txt
# Enterprise AI Task Management System Requirements
# Complete technology stack for production deployment

# =============================================================================
# 🔄 Data Pipeline Components
# =============================================================================

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# FastAPI for data pipeline and API serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Airflow for workflow orchestration
apache-airflow>=2.7.0
apache-airflow-providers-postgres>=5.7.0
apache-airflow-providers-http>=4.7.0

# =============================================================================
# 🧠 Model Training Components
# =============================================================================

# Core ML libraries
scikit-learn>=1.3.0
xgboost>=1.6.0

# Deep Learning frameworks
torch>=1.13.0
transformers>=4.20.0
tokenizers>=0.13.0

# TensorFlow for serving (optional)
tensorflow>=2.13.0
tensorflow-serving-api>=2.13.0

# =============================================================================
# 🔍 NLP Model Components
# =============================================================================

# BERT and transformer models
transformers[torch]>=4.20.0
sentence-transformers>=2.2.0

# Text processing
nltk>=3.8.0
spacy>=3.5.0

# TF-IDF and text features
scikit-learn>=1.3.0

# =============================================================================
# 🚀 Model Serving Components
# =============================================================================

# FastAPI for API serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# TensorFlow Serving (alternative to FastAPI)
tensorflow-serving-api>=2.13.0

# Model serialization
joblib>=1.2.0
pickle5>=0.0.11

# =============================================================================
# 📦 Database Components
# =============================================================================

# PostgreSQL
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Database connection pooling
psycopg2-pool>=1.1.0

# =============================================================================
# 🔧 Additional Enterprise Components
# =============================================================================

# HTTP requests for API calls
requests>=2.31.0

# JSON handling
jsonschema>=4.17.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Configuration management
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# =============================================================================
# 📊 Monitoring and Observability
# =============================================================================

# Metrics and monitoring
prometheus-client>=0.17.0
grafana-api>=1.0.3

# Logging
structlog>=23.1.0
python-json-logger>=2.0.0

# =============================================================================
# 🔐 Security Components
# =============================================================================

# Authentication and authorization
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTPS and SSL
cryptography>=41.0.0

# =============================================================================
# 🚀 Deployment Components
# =============================================================================

# Containerization
docker>=6.1.0

# Process management
gunicorn>=21.2.0

# Environment management
python-dotenv>=1.0.0

# =============================================================================
# 📈 Performance Optimization
# =============================================================================

# Async support
aiofiles>=23.1.0
asyncio-mqtt>=0.16.0

# Caching
redis>=4.5.0
aioredis>=2.0.0

# =============================================================================
# 🔄 Data Processing Extensions
# =============================================================================

# Data validation
marshmallow>=3.19.0
cerberus>=1.3.0

# Data serialization
msgpack>=1.0.5
orjson>=3.8.0

# =============================================================================
# 🧪 Testing and Quality Assurance
# =============================================================================

# Unit testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Integration testing
httpx>=0.24.0
testcontainers>=3.7.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# =============================================================================
# 📚 Documentation
# =============================================================================

# API documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# =============================================================================
# 🔧 Development Tools
# =============================================================================

# Development server
watchdog>=3.0.0

# Debugging
ipython>=8.12.0
jupyter>=1.0.0

# =============================================================================
# 🚀 Production Deployment
# =============================================================================

# Process management
supervisor>=4.2.5

# Load balancing
nginx>=1.24.0

# Monitoring
sentry-sdk>=1.28.0

# =============================================================================
# 📊 Optional: GPU Support
# =============================================================================

# Uncomment for GPU support
# torch>=1.13.0+cu116 -f https://download.pytorch.org/whl/torch_stable.html
# tensorflow-gpu>=2.13.0

# =============================================================================
# 🔧 Optional: Additional ML Libraries
# =============================================================================

# Advanced ML algorithms
lightgbm>=4.0.0
catboost>=1.2.0

# Hyperparameter optimization
optuna>=3.2.0
hyperopt>=0.2.7

# Feature selection
feature-engine>=1.6.0

# =============================================================================
# 📈 Optional: Advanced Monitoring
# =============================================================================

# Distributed tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Metrics collection
statsd>=4.0.1

# =============================================================================
# 🔐 Optional: Enterprise Security
# =============================================================================

# LDAP integration
python-ldap>=3.4.0

# SAML authentication
python3-saml>=1.15.0

# =============================================================================
# 📊 Optional: Advanced Analytics
# =============================================================================

# Data visualization
plotly>=5.15.0
dash>=2.11.0

# Statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0

# =============================================================================
# 🚀 Optional: Cloud Deployment
# =============================================================================

# AWS integration
boto3>=1.28.0
botocore>=1.31.0

# Google Cloud integration
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0

# Azure integration
azure-storage-blob>=12.17.0
azure-identity>=1.13.0

# =============================================================================
# 📋 Installation Notes
# =============================================================================

# Install with: pip install -r requirements_enterprise.txt
# 
# For GPU support, uncomment the GPU-specific packages above
# For cloud deployment, install the relevant cloud packages
# For production, consider using a virtual environment:
# python -m venv venv
# source venv/bin/activate  # On Windows: venv\Scripts\activate
# pip install -r requirements_enterprise.txt 